Put Experience on height-aware slider in 3 parts
I changed that awful button-back-button way of browsing the Experience
page into using the slider technology developed for Pricing. In order to
make it fit well, I combined "Hardware Troubleshooting" with "Hardware
Setup" and "Software Troubleshooting" with "Software Setup." This means
there are now three categories: "Hardware", "Software", and
"Development".

I am considering moving Development completely out of the page.

This required refactoring Pricing's code. I moved the JavaScript that
powers the slider on Pricing into it's own file under static. I then
moved the CSS (which includes functionality) into it.css. TODO: I need
to determine if this should instead be in a universal CSS file. We'll
see.

I also added support for height changing on the slider (animated!),
which was necessary because the sections differ so much. This was not
such a super significant change, but while I was doing it, I PORTED
SLIDER.JS TO JQUERY. There was so much DOM stuff it was horrible, so
that cleaned it up.